fix: tests and rename contact-job-error
