+++ hrm-service/src/contact-job/contact-job-error.ts
